// src/app/modules/barbara-damage-viewer/helpers/sensorai-parts.enum.ts
export enum SensoraiParts {

    'front_bumper',
    'left_front_wing',
    'right_tail_light',
    'left_front_tire',
    'rear_bumper',
    'right_front_wing',
    'left_foglight',
    'left_rear_tire',
    'left_light',
    'left_rear_wing',
    'right_foglight',
    'right_front_tire',
    'right_light',
    'right_rear_wing',
    'grates',
    'right_rear_tire',
    'grille',
    'decklid',
    'right_front_window',
    'doorknob',
    'hood',
    'front_windshield',
    'left_front_window',
    'wheelbrow',
    'left_front_door',
    'rear_windshield',
    'right_rear_window',
    'logo',
    'right_front_door',
    'left_mirror',
    'left_rear_window',
    'plate',
    'left_rear_door',
    'right_mirror',
    'right_doorsill',
    'right_rear_door',
    'left_tail_light',
    'left_doorsill',

    FRONT_BUMPER = 'front_bumper',
    LEFT_FRONT_WING = 'left_front_wing',
    RIGHT_TAIL_LIGHT = 'right_tail_light',
    LEFT_FRONT_TIRE = 'left_front_tire',
    REAR_BUMPER = 'rear_bumper',
    RIGHT_FRONT_WING = 'right_front_wing',
    LEFT_FOGLIGHT = 'left_foglight',
    LEFT_REAR_TIRE = 'left_rear_tire',
    LEFT_LIGHT = 'left_light',
    LEFT_REAR_WING = 'left_rear_wing',
    RIGHT_FOGLIGHT = 'right_foglight',
    RIGHT_FRONT_TIRE = 'right_front_tire',
    RIGHT_LIGHT = 'right_light',
    RIGHT_REAR_WING = 'right_rear_wing',
    GRATES = 'grates',
    RIGHT_REAR_TIRE = 'right_rear_tire',
    GRILLE = 'grille',
    DECKLID = 'decklid',
    RIGHT_FRONT_WINDOW = 'right_front_window',
    DOORKNOB = 'doorknob',
    HOOD = 'hood',
    FRONT_WINDSHIELD = 'front_windshield',
    LEFT_FRONT_WINDOW = 'left_front_window',
    WHEELBROW = 'wheelbrow',
    LEFT_FRONT_DOOR = 'left_front_door',
    REAR_WINDSHIELD = 'rear_windshield',
    RIGHT_REAR_WINDOW = 'right_rear_window',
    LOGO = 'logo',
    RIGHT_FRONT_DOOR = 'right_front_door',
    LEFT_MIRROR = 'left_mirror',
    LEFT_REAR_WINDOW = 'left_rear_window',
    PLATE = 'plate',
    LEFT_REAR_DOOR = 'left_rear_door',
    RIGHT_MIRROR = 'right_mirror',
    RIGHT_DOORSILL = 'right_doorsill',
    RIGHT_REAR_DOOR = 'right_rear_door',
    LEFT_TAIL_LIGHT = 'left_tail_light',
    LEFT_DOORSILL = 'left_doorsill'

}